feat: 그룹 기록함 하위 페이지는 모두 그룹 기록함 오픈그래프 적용되도록 수정(#192)
- layout에 오픈그래프 적용해 하위 페이지에서 공통되게 적용되도록 수정
